fix(autolykos): correct AMD DAG/search 32-bit index overflow and propagate hostHeader
Two AMD-resolver bugs made the miner emit non-solutions the pool rejected as
"Share above target", then drop the genuine solutions before submit:

1. 32-bit overflow in the DAG byte index. autolykos_v2_dag.cl zeroed each
   element's last byte via dag[tid*32+31] and autolykos_v2_search.cl read the
   lookup via dag[h3*32+i]; both compute the byte offset in uint, which wraps for
   index >= 2^27 (the 4 GiB boundary). On a live mainnet table (N = 216,430,305,
   ~6.9 GB) every element above 4 GiB kept a non-zero top byte and search read the
   wrong element, corrupting the verify sum and the hit. Cast the index to ulong in
   both kernels. The CUDA path has the same dag.cuh overflow (luminousmining/miner#159).

2. parameters.hostHeader was never assigned in the AMD resolver (the NVIDIA resolver
   sets it), so isValidShare() in getResultCache hashed GPU candidates against a zero
   header and rejected every real share. Set it in updateConstants.

Tests: pin the AMD kernels to the canonical Ergo height-614400 vector via the raw
verify-result buffer; dagFullyGeneratedAtLiveHeight1803848 reads back DAG elements
straddling the 4 GiB boundary against canonical genElementV2 (RED on the overflow);
findNonce asserts updateConstants propagates hostHeader. Verified live on NiceHash
(RX 9070 XT): 3 accepted shares, 0 rejected.

+++ sources/algo/autolykos/opencl/autolykos_v2_dag.cl
@@ -79,5 +79,10 @@ void autolykos_v2_build_dag(
     {
         ((__global ulong *)dag)[(tid + 1) * 4 - i - 1] = as_ulong(as_uchar8(h[i]).s76543210);
     }
-    ((__global uchar *)dag)[tid * 32 + 31] = 0;
+    // (ulong) cast is required: tid*32 overflows 32-bit for tid >= 2^27 (the 4 GiB
+    // byte boundary), which silently retargets this zeroing write to a low element
+    // and leaves the high elements' top byte = H[0] instead of 0 -- corrupting the
+    // verify kernel's element sum. See ResolverAutolykosv2AmdTest.dagFullyGeneratedAtLiveHeight1803848.
+    // The CUDA path (cuda/dag.cuh) has the same overflow: luminousmining/miner#159.
+    ((__global uchar *)dag)[(ulong)tid * 32 + 31] = 0;
 }
